Fix signed vs unsigned issues, fix non-Windows code included everywhere, fix path too

+++ rcxspy/rcxspy.cpp
@@ -168,7 +168,7 @@ void Reader::continueDownload()
 
 	lastBlock = block;
 
-	for (int i=0; i<length && codePos < codeLength; ++i) {
+	for (ULong i=0; i<length && codePos < codeLength; ++i) {
 		code[codePos++] = packet[5 + i];
 	}
